requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-autoslug==1.9.7
django-mysql==3.6.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
